+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # develop
+  * Feature: added the ability to pass :remote_sip_address to #invite, a remote address to use in place of the SIPp constructed one
 
 # [0.6.0](https://github.com/mojolingo/sippy_cup/compare/v0.5.0...v0.6.0)
   * Change: Call limits (`number_of_calls`, `concurrent_max` and `calls_per_second`) no longer have default values for simplicity of UAS scenarios. The value of `to_user` now defaults to the SIPp default of `s`.